Started with the @lit-labs/ssr renderer
