+++ packages/cli/package.json
@@ -24,6 +24,7 @@
     "@babel/plugin-proposal-export-default-from": "^7.10.4",
     "@babel/plugin-proposal-nullish-coalescing-operator": "^7.12.1",
     "@babel/plugin-proposal-optional-chaining": "^7.12.7",
+    "@babel/plugin-transform-modules-commonjs": "^7.14.5",
     "@babel/preset-env": "^7.11.5",
     "@babel/register": "^7.13.16",
     "@github-docs/frontmatter": "^1.3.1",
