+++ packages/luna/src/framework/bootstrap.js
@@ -3,12 +3,9 @@ dotenv.config();
 
 // TODO: allow browser globals as externals
 
-global.SSR = true;
 global.HTMLElement = class {};
 
-global.window = {};
-global.document = {
-    getElementById() {},
-};
+global.SSR = true;
+
 global.CustomEvent = class {};
 global.currentWorkingDirectory = process.cwd();
