--- packages/renderer/element-js/template-renderer.js
@@ -1,9 +1,0 @@
-import {renderToString} from "@popeindustries/lit-html-server";
-
-export default class TemplateRenderer {
-    async renderToString(template) {
-        return renderToString(template, {
-            serializePropertyAttributes: true,
-        });
-    }
-}
