+++ packages/renderer/src/lit/stubs/decorators.js
@@ -1,0 +1,8 @@
+const decorator = () => {
+    return function(element) { }
+};
+
+const customElement = decorator;
+const property = decorator;
+
+export { customElement, property }
