+++ packages/renderer/src/lit/stubs/index.js
@@ -1,0 +1,4 @@
+import { html } from 'lit-html';
+const css = html;
+class LitElement {}
+export { LitElement, html, css }
