+++ packages/renderer/src/lit/template-renderer.js
@@ -1,0 +1,22 @@
+import { Readable } from 'stream';
+import { render } from '@lit-labs/ssr/lib/render-with-global-dom-shim';
+
+const streamToString = (stream) => {
+    const chunks = [];
+    return new Promise((resolve, reject) => {
+        stream.on('data', (chunk) => chunks.push(Buffer.from(chunk)));
+        stream.on('error', (err) => reject(err));
+        stream.on('end', () => resolve(Buffer.concat(chunks).toString('utf8')));
+    })
+};
+
+export default class TemplateRenderer {
+    async renderToString(template) {
+        // We need a dynamic import here, because we need to force rollup to
+        // import the render function after loading the additional imports defined
+        // in this renderer.
+
+        const stream = Readable.from(render(template, { elementRenderers: [ ], customElementHostStack: [], customElementInstanceStack: [] }));
+        return streamToString(stream);
+    }
+}
